backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-generativeai
pyjwt
cryptography
httpx
fhir.resources
